Implementation of new features in the connectivity module. Analyzer first commit.
