+++ medusa/graph_theory/__init__.py
@@ -1,0 +1,10 @@
+from medusa.graph_theory import *
+
+
+import medusa.graph_theory as mds_gt
+
+def calcualte_metric(**kwargs):
+
+    pass
+
+mds_gt.calcualte_metric()
